The first released version of my MythTV remote for Sailfish OS
This is based on my older code that was never publicly released and was
written for Nokia N9. That includes mostly just remote.cpp and remote.h
files. Most of the QML is rewritten or heavily modified for Silica.

This application can connect to Frontend and send commands and it
includes most of the controls needed to watch tv or recordings. I've
used it somewhat successfully.

Known bugs and future improvements:
 - Connect dialog doesn't slide to connect
 - Cover doesn't change icon when pausing/playing
 - Volume control is not possible ATM
 - Can't remember the last connected frontend yet
 - All buttons are text only, icons needed
 - Application icon needed
 - Will be released in Harbour later